requirements.txt
Django==4.1.6
django-cors-headers==3.14.0
djangorestframework==3.14.0
python-decouple==3.7
